+++ src/konsepy/__init__.py
@@ -1,2 +1,2 @@
 """Framework for build NLP information extraction systems using regular expressions."""
-__version__ = '0.0.11'
+__version__ = '0.0.12'