+++ CHANGELOG.md
@@ -17,9 +17,15 @@ Types of changes:
 
 ## [Unreleased]
 
+### Added
+
+* Convert jsonl file to `datasets.DatasetDict` for processing with `transformers` library
+* Function to train a model on the `datasets.DatasetDict` using a pretrained model
+
 ### Changed
 
 * `get_text_snippets` when applied to existing concepts will default to use the `RUN_REGEX_FUNC` unless the `--no-regex-func` flag is supplied
+
 
 ## [0.0.11] - 2023-11-16